Move `CommentSection` component to separate module
